+++ src/Actions/index.ts
@@ -1,8 +1,7 @@
 import axios from "axios";
 import { Dispatch } from "redux";
-import { baseurl } from "../constant";
-// remove unwanted imports
-import { Url, EditedUrl, UrlListParams } from "../types";
+import { baseurl } from "../constant/constant";
+import { Url, UrlListParams } from "../types/types";
 
 export const Update_Token_Action = function (token: string) {
   return {
@@ -11,40 +10,12 @@ export const Update_Token_Action = function (token: string) {
   };
 };
 
-// if action is returning an object, name should be all capital, e.g. Update_Urls_List => UPDATE_URLS_LIST
-export const Update_Urls_List = function (urlsList: object[]) {
+export const UPDATE_URLS_LIST = function (urlsList: object[]) {
   return {
     type: "UPDATING_LIST_OF_URLS",
     payload: urlsList,
   };
 };
-
-// export function Sign_UP_Api(
-//   { email, password, fName, lName, gender },
-//   successCb
-// ) {
-//   return (dispatch) => {
-//     axios
-//       .post(`${baseurl}/signup`, {
-//         email,
-//         password,
-//         first_name: fName,
-//         last_name: lName,
-//         gender,
-//       })
-//       .then((response) => {
-//         //store.dispatch({
-//         // type: "UPDATE_THE_TOKEN",
-//         //    payload: ""
-//         // }) please can i keep them
-//         const action = Update_Token_Action(response.data.token);
-//         dispatch(action);
-
-//         successCb();
-//       })
-//       .catch((error) => console.error(error));
-//   };
-// }
 
 type RegistrationParams = {
   email: string;
@@ -54,7 +25,7 @@ type RegistrationParams = {
   gender: "M" | "F";
 };
 
-export function Sign_UP_Api(
+export function signup(
   { email, password, fName, lName, gender }: RegistrationParams,
   successCb: () => void
 ) {
@@ -93,33 +64,7 @@ export function EDIT_URL({
   };
 }
 
-// export function On_Edit_Api({ selectedUrl, editedUrl, token }, onClosePopup) {
-//   return (dispatch) => {
-//     axios
-//       .put(
-//         `${baseurl}/url/${selectedUrl.id}`,
-//         { full_url: editedUrl },
-//         {
-//           headers: {
-//             Authorization: `Bearer ${token}`,
-//           },
-//         }
-//       )
-//       .then((response) => {
-//         const action = EDIT_URL({ editedUrl, selectedUrl });
-//         dispatch(action);
-//       })
-//       .catch((error) => {
-//         console.log(error);
-//       })
-
-//       .finally(() => {
-//         onClosePopup();
-//       });
-//   };
-// }
-
-export function Edit_Url_Api(
+export function editUrl(
   { selectedUrl, editedUrl, token }: UrlListParams,
   onClosePopup: () => void
 ) {
@@ -145,13 +90,12 @@ export function Edit_Url_Api(
   };
 }
 
-// all type names should start with capital letter
-type deleteParams = {
+type DeleteParams = {
   id: string;
   token: string;
   urlsList: Url[];
 };
-export function On_Delete_Api({ id, token, urlsList }: deleteParams) {
+export function On_Delete_Api({ id, token, urlsList }: DeleteParams) {
   return async (dispatch: Dispatch) => {
     try {
       await axios.delete(`${baseurl}/url/${id}`, {
@@ -160,7 +104,7 @@ export function On_Delete_Api({ id, token, urlsList }: deleteParams) {
         },
       });
       dispatch(
-        Update_Urls_List(
+        UPDATE_URLS_LIST(
           urlsList.filter((el) => {
             return el.id !== id;
           })
@@ -172,24 +116,7 @@ export function On_Delete_Api({ id, token, urlsList }: deleteParams) {
   };
 }
 
-// export function Use_Effect_Api({ token }) {
-//   console.log("useeffect run");
-//   return (dispatch) => {
-//     axios
-//       .get(`${baseurl}/urls`, {
-//         headers: {
-//           Authorization: `Bearer ${token}`,
-//         },
-//       })
-
-//       .then((response) => {
-//         dispatch(Update_Urls_List(response.data));
-//       })
-//       .catch((error) => console.error(error));
-//   };
-// }
-
-export function List_Of_Urls({ token }: { token: string }) {
+export function getUrlsList({ token }: { token: string }) {
   return async (dispatch: Dispatch) => {
     try {
       const response = await axios.get(`${baseurl}/urls`, {
@@ -198,19 +125,19 @@ export function List_Of_Urls({ token }: { token: string }) {
         },
       });
 
-      dispatch(Update_Urls_List(response.data));
+      dispatch(UPDATE_URLS_LIST(response.data));
     } catch (error) {
       console.error(error);
     }
   };
 }
 
-type inputParams = {
+type InputParams = {
   token: string;
   longUrl: string;
   props: any;
 };
-export function Input_Url_Api({ token, longUrl, props }: inputParams) {
+export function Input_Url_Api({ token, longUrl, props }: InputParams) {
   console.log(props);
   return async (dispatch: Dispatch) => {
     try {
@@ -230,42 +157,3 @@ export function Input_Url_Api({ token, longUrl, props }: inputParams) {
     }
   };
 }
-
-// export function Input_Url_Api({ token, longUrl, props }) {
-//   return (dispatch) => {
-//     axios
-//       .post(
-//         `${baseurl}/url`,
-//         { url: longUrl },
-//         {
-//           headers: {
-//             Authorization: `Bearer ${token}`,
-//           },
-//         }
-//       )
-//       .then((response) => {
-//         props.setShortUrl(response.data.short_url);
-//       });
-//   };
-// }
-
-// export function On_Delete_Api({ id, token, urlsList }) {
-//   return (dispatch) => {
-//     axios
-//       .delete(`${baseurl}/url/${id}`, {
-//         headers: {
-//           Authorization: `Bearer ${token}`,
-//         },
-//       })
-//       .then((response) => {
-//         dispatch(
-//           Update_Urls_List(
-//             urlsList.filter((el) => {
-//               return el.id !== id;
-//             })
-//           )
-//         );
-//       })
-//       .catch((error) => console.error(error));
-//   };
-// }
