--- src/App.tsx
@@ -1,71 +1,0 @@
-// import { useState } from 'react'
-// import reactLogo from './assets/react.svg'
-// import './App.css'
-
-// function App() {
-//   const [count, setCount] = useState(0)
-
-//   return (
-//     <div className="App">
-//       <div>
-//         <a href="https://vitejs.dev" target="_blank">
-//           <img src="/vite.svg" className="logo" alt="Vite logo" />
-//         </a>
-//         <a href="https://reactjs.org" target="_blank">
-//           <img src={reactLogo} className="logo react" alt="React logo" />
-//         </a>
-//       </div>
-//       <h1>Vite + React</h1>
-//       <div className="card">
-//         <button onClick={() => setCount((count) => count + 1)}>
-//           count is {count}
-//         </button>
-//         <p>
-//           Edit <code>src/App.tsx</code> and save to test HMR
-//         </p>
-//       </div>
-//       <p className="read-the-docs">
-//         Click on the Vite and React logos to learn more
-//       </p>
-//     </div>
-//   )
-// }
-
-// export default App
-
-import React from "react";
-import { BrowserRouter, Routes, Route } from "react-router-dom";
-import ProtectedRoutes from "./ProtectedRoutes";
-import SignUp from "./SignUp";
-import ShortenUrlPage from "./ShortenUrlPage";
-import ListOfUrl from "./ListOfUrl";
-import { store } from "./Store";
-
-function App() {
-  const token = store.getState().token;
-
-  return (
-    <BrowserRouter>
-      <Routes>
-        <Route path="/signup" element={<SignUp />} />
-        <Route
-          path="/"
-          element={
-            <ProtectedRoutes>
-              <ShortenUrlPage />
-            </ProtectedRoutes>
-          }
-        />
-        <Route
-          path="/urls-list"
-          element={
-            <ProtectedRoutes>
-              <ListOfUrl />
-            </ProtectedRoutes>
-          }
-        />
-      </Routes>
-    </BrowserRouter>
-  );
-}
-export default App;
