+++ src/Store/index.ts
@@ -1,4 +1,4 @@
 import { applyMiddleware, createStore } from "redux";
-import { reducer } from "../Reducer/index";
+import { reducer } from "../reducer/index";
 import thunk from "redux-thunk";
 export const store = createStore(reducer, applyMiddleware(thunk));
